player position: int to float

+++ include/Player.h
@@ -7,7 +7,7 @@ class Player: public GameObject
 {
 public:
 
-    Player( SDL_Renderer* r , Texture* t , int x , int y );
+    Player( SDL_Renderer* r , Texture* t , float x , float y );
 
     void loadSprites();
